+++ spec/lib/ratonvirus/scanner/eicar_spec.rb
@@ -23,8 +23,12 @@ describe Ratonvirus::Scanner::Eicar do
       end
 
       context "when there is an exception" do
+        before do
+          allow(Digest::SHA256).to receive(:file).and_raise("Example")
+        end
+
         it "results to antivirus_client_error error" do
-          expect(Digest::SHA256).to receive(:file).and_raise("Example")
+          expect(Digest::SHA256).to receive(:file)
           expect(subject.virus?(path)).to be(true)
           expect(subject.errors).to contain_exactly(:antivirus_client_error)
         end
@@ -39,9 +43,22 @@ describe Ratonvirus::Scanner::Eicar do
         expect(subject.errors).to contain_exactly(:antivirus_virus_detected)
       end
 
+      context "when the test file has a newline at the end of it" do
+        let(:path) { ratonvirus_file_fixture("infected_file_newline.pdf") }
+
+        it "results to antivirus_virus_detected error" do
+          expect(subject.virus?(path)).to be(true)
+          expect(subject.errors).to contain_exactly(:antivirus_virus_detected)
+        end
+      end
+
       context "when there is an exception" do
+        before do
+          allow(Digest::SHA256).to receive(:file).and_raise("Example")
+        end
+
         it "results to antivirus_client_error error" do
-          expect(Digest::SHA256).to receive(:file).and_raise("Example")
+          expect(Digest::SHA256).to receive(:file)
           expect(subject.virus?(path)).to be(true)
           expect(subject.errors).to contain_exactly(:antivirus_client_error)
         end