+++ spec/lib/ratonvirus/storage/filepath_spec.rb
@@ -14,7 +14,8 @@ describe Ratonvirus::Storage::Filepath do
 
       context "and attribute has changed" do
         it "returns expected result" do
-          expect(record).to receive(:x_changed?).and_return(true)
+          allow(record).to receive(:x_changed?).and_return(true)
+          expect(record).to receive(:x_changed?)
           expect(record).not_to receive(:changed?)
 
           expect(subject.changed?(record, :x)).to be(true)
@@ -23,7 +24,8 @@ describe Ratonvirus::Storage::Filepath do
 
       context "and attribute has not changed" do
         it "returns expected result" do
-          expect(record).to receive(:x_changed?).and_return(false)
+          allow(record).to receive(:x_changed?).and_return(false)
+          expect(record).to receive(:x_changed?)
           expect(record).not_to receive(:changed?)
 
           expect(subject.changed?(record, :x)).to be(false)
@@ -153,18 +155,21 @@ describe Ratonvirus::Storage::Filepath do
 
       context "with asset responding true to .empty?" do
         before do
-          expect(asset).to receive(:empty?).and_return(true)
+          allow(asset).to receive(:respond_to?).with(:empty?).and_return(true)
+          expect(asset).to receive(:respond_to?).with(:empty?)
+          allow(asset).to receive(:empty?).and_return(true)
         end
 
         it "does not call asset.respond_to?" do
-          expect(asset).not_to receive(:respond_to?)
+          expect(asset).not_to receive(:respond_to?).with(:path)
           expect { |b| subject.asset_path(asset, &b) }.not_to yield_control
         end
       end
 
       context "with asset responding false to .empty?" do
         before do
-          expect(asset).to receive(:empty?).and_return(false)
+          allow(asset).to receive(:empty?).and_return(false)
+          expect(asset).to receive(:empty?)
         end
 
         context "and asset not responding to .path" do
@@ -178,14 +183,31 @@ describe Ratonvirus::Storage::Filepath do
         context "and asset responding to .path" do
           let(:path) { double }
 
+          before do
+            allow(asset).to receive(:path).and_return(path)
+            expect(asset).to receive(:path)
+          end
+
           it "yields with the response of asset.path" do
             # Call to asset.path
-            expect(asset).to receive(:path).and_return(path)
-            expect(path).to receive(:empty?).and_return(false)
-
+            allow(path).to receive(:empty?).and_return(false)
+            expect(path).to receive(:empty?)
             expect { |b| subject.asset_path(asset, &b) }.to(
               yield_with_args(path)
             )
+          end
+
+          context "when path is empty" do
+            before do
+              allow(path).to receive(:empty?).and_return(true)
+              expect(path).to receive(:empty?)
+            end
+
+            it "does not yields the asset" do
+              expect { |b| subject.asset_path(asset, &b) }.not_to(
+                yield_with_args(asset)
+              )
+            end
           end
         end
       end
@@ -196,13 +218,15 @@ describe Ratonvirus::Storage::Filepath do
     let(:asset) { double }
 
     it "does not remove the asset when it is not a file" do
-      expect(File).to receive(:file?).and_return(false)
+      allow(File).to receive(:file?).and_return(false)
+      expect(File).to receive(:file?)
       expect(FileUtils).not_to receive(:remove_file)
       subject.asset_remove(asset)
     end
 
     it "removes the asset when it is a file" do
-      expect(File).to receive(:file?).and_return(true)
+      allow(File).to receive(:file?).and_return(true)
+      expect(File).to receive(:file?)
       expect(FileUtils).to receive(:remove_file).with(asset)
       subject.asset_remove(asset)
     end