+++ spec/spec_helper_rails.rb
@@ -4,11 +4,10 @@
 require "active_model"
 require "active_storage"
 require_relative "../app/validators/antivirus_validator"
-require "tempfile"
 
 # Default spec helper
 require "spec_helper"
 
 # Add the translations load path for I18n
 base_path = File.expand_path(File.dirname(File.dirname(__FILE__)))
-I18n.load_path << Dir[base_path + "/config/locales/*.yml"]
+I18n.load_path << Dir["#{base_path}/config/locales/*.yml"]